LayoutTests/storage/indexeddb/modern/index-get-count-basic.html

This tests the most basic operation of the IDBIndex methods get(), getKey(), and count().<br>
It doesn't actually do anything other than exercise the requests themselves because we don't actually index yet.<br>
<div id="logger"></div>
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function done()
{
    alert("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

function gol(message)
{
    document.getElementById("logger").innerHTML += message + "<br>";
}

var createRequest = window.indexedDB.open("IndexGetCountBasicDatabase", 1);

createRequest.onupgradeneeded = function(event) {
    alert("Initial upgrade needed: Old version - " + event.oldVersion + " New version - " + event.newVersion);

    var versionTransaction = createRequest.transaction;
    var database = event.target.result;
    var objectStore = database.createObjectStore("TestObjectStore");
    var index = objectStore.createIndex("TestIndex", "foo");

    objectStore.put("bar", "foo");
    
    var request1 = index.get("bar");
    request1.onsuccess = function() {
        alert("get result is: " + request1.result);
    }

    request2 = index.getKey("bar");
    request2.onsuccess = function() {
        alert("getKey result is: " + request2.result);
    }
    
    var request3 = index.count();
    request3.onsuccess = function() {
        alert("count result is: " + request3.result);
    }
        
    versionTransaction.onabort = function(event) {
        alert("Initial upgrade versionchange transaction unexpected aborted");
        done();
    }

    versionTransaction.oncomplete = function(event) {
        alert("Initial upgrade versionchange transaction complete");
        done();
    }

    versionTransaction.onerror = function(event) {
        alert("Initial upgrade versionchange transaction unexpected error" + event);
        done();
    }
}

</script>
